refactor(checker): split call diagnostic helpers
